Se logra crear funcionalidad para el botón de like y el contador, pero aún falla puesto que toca recargar la página para ver los posts y en la página principal aparecen los posts, se empieza a arreglar el CSS

+++ src/view/wall.js
@@ -90,6 +90,7 @@ export default () => {
        }
        querySnapshot.forEach((doc) => {
            const post = doc.data();
+           console.log(post.likes)
            console.log('post: ', post); 
            paintPost += `
            <div class="postPublications">
@@ -111,7 +112,9 @@ export default () => {
        const btnsDelete = comentContainer.querySelectorAll('.btn-delete')
        btnsDelete.forEach(btn => {
            btn.addEventListener('click', ({target: { dataset }}) => {
-               deleteComent(dataset.id)
+            if (window.confirm('¿Estás segur@ de borrar este post?')) {   
+            deleteComent(dataset.id);
+            }
            })
        })
 
@@ -134,7 +137,7 @@ export default () => {
        like.forEach((btn) => {
            btn.addEventListener('click', (e) => {
                const userId = auth.currentUser.uid;
-               updateLikeBtn(e.target.value, userId);
+               updateLikeBtn(e.target.dataset.id, userId);
            });
        });
